@@ -10,16 +10,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.3.0",
+    "@angular/animations": "~14.2.7",
     "@angular/cdk": "^13.3.2",
-    "@angular/common": "~13.3.0",
-    "@angular/compiler": "~13.3.0",
-    "@angular/core": "~13.3.0",
-    "@angular/forms": "~13.3.0",
+    "@angular/common": "~14.2.7",
+    "@angular/compiler": "~14.2.7",
+    "@angular/core": "~14.2.7",
+    "@angular/forms": "~14.2.7",
     "@angular/material": "^13.3.2",
-    "@angular/platform-browser": "~13.3.0",
-    "@angular/platform-browser-dynamic": "~13.3.0",
-    "@angular/router": "~13.3.0",
+    "@angular/platform-browser": "~14.2.7",
+    "@angular/platform-browser-dynamic": "~14.2.7",
+    "@angular/router": "~14.2.7",
     "angular-material": "^1.2.5",
     "moment": "^2.29.4",
     "ngx-owl-carousel-o": "^7.0.2",
@@ -30,7 +30,7 @@
   "devDependencies": {
     "@angular-devkit/build-angular": "~13.3.1",
     "@angular/cli": "^14.2.6",
-    "@angular/compiler-cli": "~13.3.0",
+    "@angular/compiler-cli": "~14.2.7",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
     "jasmine-core": "~4.0.0",
